capturing progress
